+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## master
+## 10.1.10 Oct 7, 2022
 
 Changes: